extremely hacky direction-neuron-to-direction-neuron connections put in place

+++ EyeBotReboot/EyeBotReboot/Components/Axons/ReceptorAxon.cs
@@ -25,7 +25,7 @@ namespace EyeBotReboot.Components.Axons
         }
 
         public ILaplaceDendrite Dendrite { get; set; }
-        public float BrightnessCharge { get; set; }
+        public double BrightnessCharge { get; set; }
 
         public void Fire()
         {
