+++ EyeBotReboot/EyeBotReboot/Components/Muscle/EyeMuscle.cs
@@ -8,7 +8,7 @@ namespace EyeBotReboot.Components.Muscle
 {
     public class EyeMuscle: INeuron
     {
-        public float Charge { get; set; }
+        public double Charge { get; set; }
         public ReceptorField ReceptorField { get; set; } //should this be an axon->dendrite connection instead?  Probably, but not sure, since the muscle actually, literally, moves the eye causing its relationship with the environment to change instead of just executing internal state changes through charge-exchange.  Need to look into more
         public int EnvironmentFocusX { get; set; }
         public int EnvironmentFocusY { get; set; }
